feat: move URL Parser to foundation.
